Organize tests into module-based structure
